requirements.txt
# Add your project dependencies here
redis==5.0.1
redis-timeseries==0.1.9
python-dateutil==2.8.2
schedule==1.2.0
requests==2.31.0
pytest==7.4.4
python-dotenv==1.0.0
